When rendering the LightboxView try to use a Photo in the Photos list.
In the case where the Photos list data is cached, this attempts to use
the Photo instance that is already in the list so the prev/next links
can be added.

This is not totally correct yet, because we want to load more photos
like the GirdView does when you're viewing the last photo in the list.