Shared templates are always loaded from filesystem in dev mode.
When in production mode, the templates will only be read from the
filesystem once. To aid development, this changes causes the shared
templates to always be read from the filesystem. This way a change can
be saved to the file, and the reload the browser to see it without
having to restart the server.